feat: add metadata title to signIn, signUp and cart page
